Add support for OCI Client metrics
