+++ cloud/metrics/http_request_dispatcher_wrapper.go
@@ -1,0 +1,48 @@
+/*
+Copyright (c) 2023 Oracle and/or its affiliates.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package metrics
+
+import (
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/oracle/oci-go-sdk/v65/common"
+)
+
+type DispatcherWrapper struct {
+	dispatcher common.HTTPRequestDispatcher
+	region     string
+}
+
+func (wrapper DispatcherWrapper) Do(req *http.Request) (*http.Response, error) {
+	service := strings.Split(req.URL.Path, "/")[2]
+	t := time.Now()
+	resp, err := wrapper.dispatcher.Do(req)
+	defer func() {
+		IncRequestCounter(err, service, req.Method, wrapper.region, resp)
+		ObserverRequestDuration(service, req.Method, wrapper.region, time.Since(t))
+	}()
+	return resp, err
+}
+
+func NewDispatcherWrapper(wrapper common.HTTPRequestDispatcher, region string) DispatcherWrapper {
+	return DispatcherWrapper{
+		dispatcher: wrapper,
+		region:     region,
+	}
+}
